Resolved Conflicts:
	wrappers/csharp/Intel.RealSense/CMakeLists.txt
	wrappers/csharp/Intel.RealSense/Devices/AdvancedDevice.cs
	wrappers/csharp/Intel.RealSense/Devices/Device.cs
	wrappers/csharp/Intel.RealSense/Devices/DeviceList.cs
	wrappers/csharp/Intel.RealSense/Devices/PlaybackDevice.cs
	wrappers/csharp/Intel.RealSense/Devices/RecordDevice.cs
	wrappers/csharp/Intel.RealSense/Devices/SoftwareDevice.cs
	wrappers/csharp/Intel.RealSense/Frames/DepthFrame.cs
	wrappers/csharp/Intel.RealSense/Frames/Frame.cs
	wrappers/csharp/Intel.RealSense/Frames/FramePool.cs
	wrappers/csharp/Intel.RealSense/Frames/FrameSet.cs
	wrappers/csharp/Intel.RealSense/Frames/Points.cs
	wrappers/csharp/Intel.RealSense/Frames/VideoFrame.cs
	wrappers/csharp/Intel.RealSense/Helpers.cs
	wrappers/csharp/Intel.RealSense/Pipeline.cs
	wrappers/csharp/Intel.RealSense/Processing/Align.cs
	wrappers/csharp/Intel.RealSense/Processing/Colorizer.cs
	wrappers/csharp/Intel.RealSense/Processing/CustomProcessingBlock.cs
	wrappers/csharp/Intel.RealSense/Processing/DecimationFilter.cs
	wrappers/csharp/Intel.RealSense/Processing/DisparityTransform.cs
	wrappers/csharp/Intel.RealSense/Processing/FrameSource.cs
	wrappers/csharp/Intel.RealSense/Processing/HoleFillingFilter.cs
	wrappers/csharp/Intel.RealSense/Processing/PointCloud.cs
	wrappers/csharp/Intel.RealSense/Processing/ProcessingBlock.cs
	wrappers/csharp/Intel.RealSense/Processing/SpatialFilter.cs
	wrappers/csharp/Intel.RealSense/Processing/Syncer.cs
	wrappers/csharp/Intel.RealSense/Processing/TemporalFilter.cs
	wrappers/csharp/Intel.RealSense/Sensors/Sensor.cs
	wrappers/csharp/Intel.RealSense/Sensors/SoftwareSensor.cs
	wrappers/csharp/Intel.RealSense/Types/CameraInfo.cs
	wrappers/csharp/Intel.RealSense/Types/Delegates.cs
	wrappers/csharp/Intel.RealSense/Types/Distortion.cs
	wrappers/csharp/Intel.RealSense/Types/ExceptionType.cs
	wrappers/csharp/Intel.RealSense/Types/Extension.cs
	wrappers/csharp/Intel.RealSense/Types/Extrinsics.cs
	wrappers/csharp/Intel.RealSense/Types/Format.cs
	wrappers/csharp/Intel.RealSense/Types/FrameMetadataValue.cs
	wrappers/csharp/Intel.RealSense/Types/Intrinsics.cs
	wrappers/csharp/Intel.RealSense/Types/LogSeverity.cs
	wrappers/csharp/Intel.RealSense/Types/Matchers.cs
	wrappers/csharp/Intel.RealSense/Types/NotificationCategory.cs
	wrappers/csharp/Intel.RealSense/Types/Option.cs
	wrappers/csharp/Intel.RealSense/Types/PlaybackStatus.cs
	wrappers/csharp/Intel.RealSense/Types/RecordingMode.cs
	wrappers/csharp/Intel.RealSense/Types/Rs400VisualPreset.cs
	wrappers/csharp/Intel.RealSense/Types/SoftwareVideoFrame.cs
	wrappers/csharp/Intel.RealSense/Types/SoftwareVideoStream.cs
	wrappers/csharp/Intel.RealSense/Types/Sr300VisualPreset.cs
	wrappers/csharp/Intel.RealSense/Types/Stream.cs
	wrappers/csharp/Intel.RealSense/Types/TimestampDomain.cs
	wrappers/csharp/Intel.RealSense/Types/VideoStream.cs

+++ include/librealsense2/rsutil.h
@@ -107,9 +107,9 @@ static bool is_pixel_in_line(const float curr[2], const float start[2], const fl
 static void adjust_2D_point_to_boundary(float p[2], int width, int height)
 {
     if (p[0] < 0) p[0] = 0;
-    if (p[0] > width) p[0] = width;
+    if (p[0] > width) p[0] = (float)width;
     if (p[1] < 0) p[1] = 0;
-    if (p[1] > height) p[1] = height;
+    if (p[1] > height) p[1] = (float)height;
 }
 
 /* Find projected pixel with unknown depth search along line. */
